+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/mod.rs
@@ -111,7 +111,7 @@ impl ExecuteCall for CallContractRequest {
             storage_address: contract_address,
             caller_address: TryFromHexStr::try_from_hex_str(TEST_ADDRESS).unwrap(),
             call_type: CallType::Call,
-            initial_gas: u64::MAX,
+            initial_gas: i64::MAX as u64,
         };
 
         call_entry_point(
